exec loop cleaned
