+++ Sources/Execution/execution.c
@@ -14,139 +14,18 @@
 
 extern long long g_status;
 
-char *get_exec(char *cmd)
+int	is_builtin(t_tokens *temp)
 {
-	size_t	j;
-	char	*res;
-
-	j = 0;
-	while (cmd[j])
-		j++;
-	while (j != 0 && cmd[j] != '/')
-		j--;
-	j++;
-	res = ft_substr(cmd, j, ft_strlen(cmd));
-	return (res);
+	if (!ft_strcmp(temp->cmd_arr[0], "echo") || !ft_strcmp(temp->cmd_arr[0], "cd") ||
+		!ft_strcmp(temp->cmd_arr[0], "exit") || !ft_strcmp(temp->cmd_arr[0], "env") ||
+		!ft_strcmp(temp->cmd_arr[0], "unset") || !ft_strcmp(temp->cmd_arr[0], "pwd") ||
+		!ft_strcmp(temp->cmd_arr[0], "export"))
+			return (1);
+		return (0);
 }
 
-char	*get_path(char *cmd)
+void	builtin_forwarding(t_tokens *temp, t_mshell *shell)
 {
-	size_t	i;
-	char	*res;
-
-	i = 0;
-	while (cmd[i])
-		i++;
-	while (cmd[i] != '/')
-		i--;
-	i++;
-	res = ft_substr(cmd, 0, i);
-	return (res);
-}
-
-void bin_exec(t_mshell *shell, char **cmd_arr, int fd_in, int fd_out)
-{
-	size_t	j;
-	char	*temp;
-	char	*trim;
-	char	*exec;
-	char	**exec_split;
-
-	if (fd_out != 1)
-	{
-		dup2(fd_out, STDOUT_FILENO);
-		close(fd_out);
-	}
-	if (fd_in != 0)
-	{
-		dup2(fd_in, STDIN_FILENO);
-		close(fd_in);
-	}
-	j = -1;
-	if (!get_current_location(shell))
-		return (free_struct(shell), exit(1));
-	trim = ft_strtrim(cmd_arr[0], ".");
-	temp = ft_strjoin(shell->current_loc, trim);
-	free(trim);
-	if (execve(temp, cmd_arr, shell->menvp) == -1)
-		free(temp);
-	if (ft_strchr(cmd_arr[0], '/'))
-	{
-		exec = get_exec(cmd_arr[0]);
-		exec_split = ft_split(exec, ' ');
-		temp = ft_strjoin(get_path(cmd_arr[0]), exec);
-		if (execve(temp, exec_split, shell->menvp) == -1)
-		{
-			free(exec);
-			free_arr(exec_split);
-			free(temp);
-		}
-	}
-	if (shell->paths)
-	{
-		while (shell->paths[++j])
-		{
-			temp = ft_strjoin(shell->paths[j], cmd_arr[0]);
-			if (!temp)
-				return (free_struct(shell), exit(1));
-			if (execve(temp, cmd_arr, shell->menvp) == -1)
-				free(temp);
-		}
-		show_error(cmd_arr[0], "EXEC", 1);
-	}
-	else
-		show_error(cmd_arr[0], "EXEC", 2);
-	free_struct(shell);
-	if (fd_out != 1)
-		close(fd_out);
-	if (fd_in != 0)
-		close(fd_in);
-	exit(127);
-}
-
-void	print_single_token(t_tokens *tkn)
-{
-	if (DEBUG)
-	{
-		ft_printf("content = %s\n", tkn->content);
-		ft_printf("cmd_arr :\n_________________\n");
-		print_arr(tkn->cmd_arr);
-		ft_printf("_________________\n");
-		ft_printf("fd_in = %d, fd_out = %d\n", tkn->fd_in, tkn->fd_out);
-	}
-}
-
-char *get_cmd(char *str, size_t *i)
-{
-	char *result;
-	size_t res_len;
-	size_t j;
-
-	while (str[(*i)] != ' ')
-		(*i)++;
-	if (str[(*i) + 1] == '-')
-	{
-		while (str[(*i)] != ' ')
-			(*i)++;
-	}
-	res_len = (*i) + 1;
-	result = malloc(sizeof(char) * res_len);
-	if (!result)
-		return (NULL);
-	result[res_len - 1] = '\0';
-	j = 0;
-	while (j < res_len)
-	{
-		result[j] = str[j];
-		j++;
-	}
-	return (result);
-}
-
-void	exec_forwarding(t_tokens *temp, t_mshell *shell)
-{
-	pid_t	child;
-
 	if (!ft_strcmp(temp->cmd_arr[0], "echo"))
 		echo_case(temp->cmd_arr, temp->fd_out);
 	else if (!ft_strcmp(temp->cmd_arr[0], "cd"))
@@ -161,72 +40,34 @@ void	exec_forwarding(t_tokens *temp, t_mshell *shell)
 		pwd_case(shell, temp->cmd_arr, temp->fd_out);
 	else if (!ft_strcmp(temp->cmd_arr[0], "export"))
 		export_case(shell, temp->cmd_arr, temp->fd_out);
+}
+
+void	executable(t_tokens *temp, t_mshell *shell)
+{
+	pid_t	child;
+
+	child = fork();
+	if (child == -1)
+		return (free_struct(shell), exit(2));
+	if (!child)
+		bin_exec(shell, temp->cmd_arr, temp->fd_in, temp->fd_out);
 	else
-	{
-		child = fork();
-		if (child == -1)
-			return (free_struct(shell), exit(2));
-		if (!child)
-			bin_exec(shell, temp->cmd_arr, temp->fd_in, temp->fd_out);
-		else
-			waitpid(child, (int *)&g_status, 0);
-		if (WIFEXITED(g_status))
-			g_status = WEXITSTATUS(g_status);
-		else if (WIFSIGNALED(g_status))
-			g_status = WTERMSIG(g_status);
-		if (temp->cmd_arr)
-			free_arr(temp->cmd_arr);
-		temp->cmd_arr = NULL;
-	}	
+		waitpid(child, (int *)&g_status, 0);
+	if (WIFEXITED(g_status))
+		g_status = WEXITSTATUS(g_status);
+	else if (WIFSIGNALED(g_status))
+		g_status = WTERMSIG(g_status);
+	if (temp->cmd_arr)
+		free_arr(temp->cmd_arr);
+	temp->cmd_arr = NULL;
 }
 
-int	get_final_out(t_tokens *lst)
+void	exec_forwarding(t_tokens *temp, t_mshell *shell)
 {
-	t_tokens	*temp;
-	int			result;
-
-	temp = lst;
-	while (temp)
-	{
-		if (temp->type == RCHEVRON)
-		{
-			result = open(temp->next->content, O_RDWR | O_CREAT, 0666);
-			return (result);
-		}
-		else if (temp->type == APPEND)
-		{
-			result = open(temp->next->content, O_RDWR | O_APPEND, 0666);
-			return (result);
-		}
-		temp = temp->next;
-	}
-	return (1);
-}
-
-void	free_single_token(t_tokens *node)
-{
-	free(node->content);
-	free(node);
-}
-
-void	remove_next_token(t_tokens **lst)
-{
-	t_tokens *next;
-
-	if ((*lst)->next)
-		next = (*lst)->next->next;
+	if (is_builtin(temp))
+		builtin_forwarding(temp, shell);
 	else
-		next = NULL;
-	if ((*lst)->next && (*lst)->next->next)
-		(*lst)->next->next->prev = *lst;
-	free_single_token((*lst)->next);
-	(*lst)->next = next;
-}
-
-void	remove_heredoc_tkn(t_tokens **lst)
-{
-	while ((*lst)->next && ((*lst)->next->type == HEREDOC || (*lst)->next->type == HEREDEL))
-		remove_next_token(lst);
+		executable(temp, shell);
 }
 
 void	execution(t_mshell *shell)
@@ -238,10 +79,7 @@ void	execution(t_mshell *shell)
 	while (temp)
 	{
 		if (temp->next && temp->next->type == PIPE)
-		{
-			// print_single_token(temp);
 			handle_pipes(shell, &temp, &temp->fd_in, &temp->fd_out);
-		}
 		else
 			exec_forwarding(temp, shell);
 		temp = temp->next;
